Cleaned up csv parsing functions in their own file
